pkg/gltfv/pyproject.toml
[tool.setuptools]
packages = ["gltfv"]

[project]
name = "gltfv"
version = "0.1.0"
description = 'Crunge WebGPU/GLTF Demo'
requires-python = ">=3.8"
authors = [{ name = "Kurtis Fields", email = "kurtisfields@gmail.com" }]

dependencies = [
    "loguru ==0.7.2",
    "click ==8.1.7",
    "glfw ==2.6.4",
    "PyGLM ==2.7.1",
    "numpy ==1.26.3",
]

[project.scripts]
gltfv = 'gltfv.cli:cli'
